requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
clip==1.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
ftfy==6.3.1
hydra-core==1.3.2
iopath==0.1.10
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.4
pillow==11.3.0
pip==25.1
portalocker==3.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.2
regex==2025.7.34
rich==14.1.0
SAM-2==1.0
setuptools==78.1.1
six==1.17.0
stack-data==0.2.0
sympy==1.14.0
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
wheel==0.45.1
